+++ back/app/services/idea_custom_fields_service.rb
@@ -46,7 +46,18 @@ class IdeaCustomFieldsService
 
   attr_reader :custom_form
 
+  def native_survey?
+    return false unless custom_form
+
+    participation_context = ::ParticipationContextService.new.get_participation_context(custom_form.project)
+    return false unless participation_context
+
+    participation_context.native_survey?
+  end
+
   def default_fields
+    return [] if native_survey?
+
     ml_s = MultilocService.new
     [
       CustomField.new(