+++ front/app/containers/Admin/routes.tsx
@@ -11,6 +11,7 @@ import settingsRoutes from './settings/routes';
 import pagesRoutes from './pages/routes';
 import createAdminMessagingRoutes from './messaging/routes';
 import ideasRoutes from './ideas/routes';
+import pagesAndMenuRoutes from './pagesAndMenu/routes';
 
 // components
 import PageLoading from 'components/UI/PageLoading';
@@ -114,6 +115,7 @@ const createAdminRoutes = () => {
       createAdminProjectsRoutes(),
       settingsRoutes(),
       pagesRoutes(),
+      pagesAndMenuRoutes(),
       invitationsRoutes(),
       createAdminMessagingRoutes(),
       ideasRoutes(),