+++ CHANGELOG.md
@@ -1,8 +1,66 @@
 # Changelog
 
-### Next release
+## Next release
+
+### Added
+
+- [CL-1097] Added the ability for users to search for projects and folders by keyword on the main projects index page
+
+### Changed
+
+- Improve error handling and user feedback in the forms for creating pages and custom navigation items
+
+## 2022-08-11
+
+### Fixed
+
+- [CL-1302] Added topics column to bulk idea import example sheet
+
+## 2022-08-09
+
+### Fixed
+
+- [CL-1289] Fix 'Submit base data' button width (representation dashboard)
+
+### Added
+
+- [CL-1273] Enable age graphs in representation dashboard
+
+## 2022-08-04
+
+### Added
+
+-[CL-5] Customizable Pages iteration 1
+
+## 2022-08-03
+
+### Added
+
+- [CL-1189] Added Turkish locale to platform
+
+## 2022-07-28_2
+
+### Fixed
+
+- [CL-1256] Fix verification using Belgian eID or Itsme application
+
+## 2022-07-28
+
+### Added
+
+- [CL-1118] Native survey feature
+- [CL-1128] Feature to bulk import ideas from an XLSX sheet
+
+## 2022-07-22
+
+### Fixed
+
+- [CL-1216] Fix slow insights export
+
+### Changed
 
 - [CL-1205] Change see less copy to read less in read more on project info and phase description
+- [CL-1140] Show sign up modal when a logged user clicks the take survey button
 
 ## 2022-07-19
 
@@ -39,6 +97,7 @@
 
 ### Added
 
+- [CL-1058] - Add desktop preview in content builder
 - [CL-1096] Add representativeness score to header of representativeness chart card
 
 ## 2022-07-07
@@ -71,6 +130,7 @@
 ### Fixed
 
 - [CL-1083] XLSX export issues with custom fields
+
 
 ## 2022-07-04